requirements.txt
numpy == 2.4.2
pandas == 3.0.1
scipy == 1.17.1
scikit-learn == 1.8.0
matplotlib == 3.10.8
seaborn == 0.13.2
jupyterlab == 4.5.5
python-dotenv == 1.2.2
matplotlib == 3.10.8
tabulate == 0.10.0